Frontend page built, javascript started, css layout done
Okay so in this project the browser webpage was built. The JavaScript was also begun in the java script the names and ideas of the flavors was made to match what the HTML page had for consistency.  The result will be aimed to be displayed in the results box so the user can see what the webpage reads their order as.